+++ jscomp/test/escape_esmodule.js
@@ -2,11 +2,11 @@
 'use strict';
 
 
-let $$__esModule = false;
+let __esModule = false;
 
 let $$default = 4;
 
-exports.$$__esModule = $$__esModule;
+exports.__esModule = __esModule;
 exports.default = $$default;
 exports.__esModule = true;
 /* No side effect */