+++ jscomp/test/global_mangles.js
@@ -6,42 +6,42 @@ let $$__dirname = 1;
 
 let $$__filename = 2;
 
-let $$clearImmediate = 3;
+let clearImmediate = 3;
 
-let $$clearInterval = 4;
+let clearInterval = 4;
 
-let $$clearTimeout = 5;
+let clearTimeout = 5;
 
-let $$console = 6;
+let console = 6;
 
 let $$exports = 7;
 
-let $$global = 8;
+let global = 8;
 
 let _module = 9;
 
-let $$process = 10;
+let process = 10;
 
 let $$require = 11;
 
-let $$setImmediate = 12;
+let setImmediate = 12;
 
-let $$setInterval = 13;
+let setInterval = 13;
 
-let $$setTimeout = 14;
+let setTimeout = 14;
 
 exports.$$__dirname = $$__dirname;
 exports.$$__filename = $$__filename;
-exports.$$clearImmediate = $$clearImmediate;
-exports.$$clearInterval = $$clearInterval;
-exports.$$clearTimeout = $$clearTimeout;
-exports.$$console = $$console;
+exports.clearImmediate = clearImmediate;
+exports.clearInterval = clearInterval;
+exports.clearTimeout = clearTimeout;
+exports.console = console;
 exports.$$exports = $$exports;
-exports.$$global = $$global;
+exports.global = global;
 exports._module = _module;
-exports.$$process = $$process;
+exports.process = process;
 exports.$$require = $$require;
-exports.$$setImmediate = $$setImmediate;
-exports.$$setInterval = $$setInterval;
-exports.$$setTimeout = $$setTimeout;
+exports.setImmediate = setImmediate;
+exports.setInterval = setInterval;
+exports.setTimeout = setTimeout;
 /* No side effect */