+++ jscomp/test/key_word_property2.js
@@ -19,13 +19,13 @@ function test(p) {
 
 let $$case = Export_keyword.$$case;
 
-let $$window = Export_keyword.$$window;
+let window = Export_keyword.window;
 
 let $$switch = Export_keyword.$$switch;
 
 exports.test2 = test2;
 exports.test = test;
 exports.$$case = $$case;
-exports.$$window = $$window;
+exports.window = window;
 exports.$$switch = $$switch;
 /* No side effect */