+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Remove obsolete `@bs.open` feature. https://github.com/rescript-lang/rescript-compiler/pull/6629
 - Drop Node.js version <18 support, due to it reaching End-of-Life. https://github.com/rescript-lang/rescript-compiler/pull/6429
 - Remove deprecated -bs-super-errors option. https://github.com/rescript-lang/rescript-compiler/pull/6814
+- Some global names and old keywords are no longer prefixed. https://github.com/rescript-lang/rescript-compiler/pull/6831
 
 #### :bug: Bug Fix